refactor (image) : 이미지 도메인 구조 재구성
